Properly cleans up test dir.

+++ t/clean-up.sh
@@ -1,5 +1,18 @@
 #!/bin/bash
 
-rm -r test-users
-rm *.log
-rm peers.sqlite
+if [ -d test-users ]; then
+    rm -r test-users
+fi
+
+if [ -d log ]; then
+    [ "$(ls -A log 2> /dev/null)" != "" ] &&\
+        rm log/*
+fi
+
+if [ -f *.log ]; then
+    rm *.log
+fi
+
+if [ -f peers.sqlite ]; then
+    rm peers.sqlite
+fi
